+++ config/seata-config-servicecomb/src/main/java/io/seata/config/servicecomb/client/CommonConfiguration.java
@@ -34,6 +34,24 @@ public interface CommonConfiguration {
     String KEY_SERVICE_ENVIRONMENT = "servicecomb.service.environment";
 
     /**
+     * registry instance configuration
+     */
+    String KEY_INSTANCE_ENVIRONMENT = "servicecomb.instance.initialStatus";
+
+    String KEY_INSTANCE_PULL_INTERVAL = "servicecomb.instance.pull.interval";
+
+    String KEY_INSTANCE_HEALTH_CHECK_INTERVAL = "servicecomb.instance.healthCheck.interval";
+
+    String KEY_INSTANCE_HEALTH_CHECK_TIMES = "servicecomb.instance.healthCheck.times";
+
+    /**
+     * registry configuration
+     */
+    String KEY_REGISTRY_ADDRESS = "servicecomb.registry.address";
+
+    String KEY_REGISTRY_WATCH = "servicecomb.registry.watch";
+
+    /**
      * config center configuration
      */
     String KEY_CONFIG_ADDRESSTYPE = "servicecomb.config.type";
@@ -125,8 +143,16 @@ public interface CommonConfiguration {
     String UTF_8 = "UTF-8";
     String DEFAULT_VERSION = "1.0.0.0";
     String DEFAULT_CONFIG_URL = "http://127.0.0.1:30110";
+    String DEFAULT_REGISTRY_URL = "http://127.0.0.1:30100";
     String COMMA = ",";
     String PUBLIC = "public";
     String DEFAULT_SERVICE_POLLINGWAITSEC="10";
     String KIE = "kie";
+    String SEMICOLON = ";";
+    String COLON = ":";
+    String DEFAULT_INSTANCE_PULL_INTERVAL="15";
+    String UP = "UP";
+    String DEFAULT_INSTANCE_HEALTH_CHECK_INTERVAL = "15";
+    String DEFAULT_INSTANCE_HEALTH_CHECK_TIMES = "3";
+    String REST_PROTOCOL = "rest://";
 }
