+++ discovery/seata-discovery-servicecomb/pom.xml
@@ -45,5 +45,10 @@
             <version>1.1.1</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>seata-config-servicecomb</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 </project>
