+++ discovery/seata-discovery-servicecomb/src/main/java/io/seata/discovery/registry/servicecomb/client/ServicecombRegistryHelper.java
@@ -17,8 +17,13 @@
 package io.seata.discovery.registry.servicecomb.client;
 
 import com.google.common.eventbus.Subscribe;
-import io.seata.discovery.registry.servicecomb.client.auth.AuthHeaderProviders;
+import io.seata.config.servicecomb.client.CommonConfiguration;
+import io.seata.config.servicecomb.client.EventManager;
+import io.seata.config.servicecomb.client.auth.AuthHeaderProviders;
+import io.seata.discovery.registry.servicecomb.client.auth.RBACRequestAuthHeaderProvider;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.servicecomb.foundation.auth.AuthHeaderProvider;
+import org.apache.servicecomb.http.client.auth.RequestAuthHeaderProvider;
 import org.apache.servicecomb.http.client.common.HttpConfiguration;
 import org.apache.servicecomb.service.center.client.AddressManager;
 import org.apache.servicecomb.service.center.client.RegistrationEvents;
@@ -26,16 +31,22 @@ import org.apache.servicecomb.service.center.client.ServiceCenterClient;
 import org.apache.servicecomb.service.center.client.ServiceCenterDiscovery;
 import org.apache.servicecomb.service.center.client.ServiceCenterRegistration;
 import org.apache.servicecomb.service.center.client.ServiceCenterWatch;
+import org.apache.servicecomb.service.center.client.model.Framework;
+import org.apache.servicecomb.service.center.client.model.HealthCheck;
+import org.apache.servicecomb.service.center.client.model.HealthCheckMode;
 import org.apache.servicecomb.service.center.client.model.Microservice;
 import org.apache.servicecomb.service.center.client.model.MicroserviceInstance;
+import org.apache.servicecomb.service.center.client.model.MicroserviceInstanceStatus;
 import org.apache.servicecomb.service.center.client.model.ServiceCenterConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.InetAddress;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Locale;
 import java.util.Properties;
 
 /**
@@ -55,8 +66,6 @@ public class ServicecombRegistryHelper {
 
     private Properties properties;
 
-    private ServiceCenterConfigurationManager serviceCenterConfigurationManager;
-
     private Microservice microservice;
 
     private MicroserviceInstance microserviceInstance;
@@ -65,16 +74,17 @@ public class ServicecombRegistryHelper {
 
     private final String frameworkName;
 
+    private RequestAuthHeaderProvider requestAuthHeaderProvider;
+
     public ServicecombRegistryHelper(Properties properties,String frameworkName) {
         this.properties = properties;
         this.frameworkName = frameworkName;
-        serviceCenterConfigurationManager = new ServiceCenterConfigurationManager(properties);
     }
 
     public void register(String endPoint) throws Exception {
         initClient();
 
-        microserviceInstance = serviceCenterConfigurationManager.createMicroserviceInstance();
+        microserviceInstance = createMicroserviceInstance();
         microserviceInstance.setHostName(InetAddress.getLocalHost().getHostName());
 
         List<String> endPoints = new ArrayList<>();
@@ -92,9 +102,9 @@ public class ServicecombRegistryHelper {
         serviceCenterRegistration.startRegistration();
 
         if (CommonConfiguration.TRUE.equals(properties.getProperty(CommonConfiguration.KEY_REGISTRY_WATCH, CommonConfiguration.FALSE))) {
-            watch = new ServiceCenterWatch(serviceCenterConfigurationManager.createAddressManager(),
+            watch = new ServiceCenterWatch(createAddressManager(),
                 AuthHeaderProviders.createSslProperties(properties),
-                AuthHeaderProviders.getRequestAuthHeaderProvider(client, properties), CommonConfiguration.DEFAULT, Collections.EMPTY_MAP,
+                getRequestAuthHeaderProvider(client, properties), CommonConfiguration.DEFAULT, Collections.EMPTY_MAP,
                 EventManager.getEventBus());
         }
         EventManager.register(this);
@@ -124,12 +134,12 @@ public class ServicecombRegistryHelper {
                 if (client == null) {
 
                     try {
-                        AddressManager addressManager = serviceCenterConfigurationManager.createAddressManager();
+                        AddressManager addressManager = createAddressManager();
                         HttpConfiguration.SSLProperties sslProperties =
-                            AuthHeaderProviders.createSslProperties(properties);
+                        AuthHeaderProviders.createSslProperties(properties);
                         client = new ServiceCenterClient(addressManager, sslProperties,
-                            AuthHeaderProviders.getRequestAuthHeaderProvider(client, properties), CommonConfiguration.DEFAULT, null);
-                        microservice = serviceCenterConfigurationManager.createMicroservice(frameworkName);
+                            getRequestAuthHeaderProvider(client, properties), CommonConfiguration.DEFAULT, null);
+                        microservice = createMicroservice(frameworkName);
 
                     } catch (Exception e) {
                         throw new IllegalStateException(e);
@@ -158,5 +168,52 @@ public class ServicecombRegistryHelper {
 
     public void setEnableDiscovery(boolean enableDiscovery) {
         this.enableDiscovery = enableDiscovery;
+    }
+
+    public RequestAuthHeaderProvider getRequestAuthHeaderProvider(ServiceCenterClient client, Properties properties) {
+        if (requestAuthHeaderProvider==null){
+            List<AuthHeaderProvider> authHeaderProviders = new ArrayList<>();
+            authHeaderProviders.add(new RBACRequestAuthHeaderProvider(client, properties));
+            requestAuthHeaderProvider = AuthHeaderProviders.getRequestAuthHeaderProvider(authHeaderProviders);
+        }
+        return requestAuthHeaderProvider;
+    }
+    public Microservice createMicroservice(String frameworkName) {
+        Microservice microservice = new Microservice();
+        microservice.setAppId(properties.getProperty(CommonConfiguration.KEY_SERVICE_APPLICATION, CommonConfiguration.DEFAULT));
+        microservice.setServiceName(properties.getProperty(CommonConfiguration.KEY_SERVICE_NAME, CommonConfiguration.DEFAULT));
+        microservice.setVersion(properties.getProperty(CommonConfiguration.KEY_SERVICE_VERSION, CommonConfiguration.DEFAULT_VERSION));
+        microservice.setEnvironment(properties.getProperty(CommonConfiguration.KEY_SERVICE_ENVIRONMENT, CommonConfiguration.EMPTY));
+        Framework framework = new Framework();
+        framework.setName(frameworkName);
+        StringBuilder version = new StringBuilder();
+        version.append(frameworkName.toLowerCase(Locale.ROOT)).append(CommonConfiguration.COLON);
+        if(StringUtils.isEmpty(ServicecombRegistryHelper.class.getPackage().getImplementationVersion())){
+            version.append(ServicecombRegistryHelper.class.getPackage().getImplementationVersion());
+        }else{
+            version.append(CommonConfiguration.DEFAULT_VERSION);
+        }
+        version.append(CommonConfiguration.SEMICOLON);
+        framework.setVersion(version.toString());
+        microservice.setFramework(framework);
+        return microservice;
+    }
+
+    public MicroserviceInstance createMicroserviceInstance() {
+        MicroserviceInstance instance = new MicroserviceInstance();
+        instance.setStatus(MicroserviceInstanceStatus.valueOf(properties.getProperty(CommonConfiguration.KEY_INSTANCE_ENVIRONMENT, CommonConfiguration.UP)));
+        HealthCheck healthCheck = new HealthCheck();
+        healthCheck.setMode(HealthCheckMode.pull);
+        healthCheck.setInterval(Integer.parseInt(properties.getProperty(CommonConfiguration.KEY_INSTANCE_HEALTH_CHECK_INTERVAL, CommonConfiguration.DEFAULT_INSTANCE_HEALTH_CHECK_INTERVAL)));
+        healthCheck.setTimes(Integer.parseInt(properties.getProperty(CommonConfiguration.KEY_INSTANCE_HEALTH_CHECK_TIMES, CommonConfiguration.DEFAULT_INSTANCE_HEALTH_CHECK_TIMES)));
+        instance.setHealthCheck(healthCheck);
+        return instance;
+    }
+
+    public AddressManager createAddressManager() {
+        String address = properties.getProperty(CommonConfiguration.KEY_REGISTRY_ADDRESS, CommonConfiguration.DEFAULT_REGISTRY_URL);
+        String project = properties.getProperty(CommonConfiguration.KEY_SERVICE_PROJECT, CommonConfiguration.DEFAULT);
+        LOGGER.info("Using service center, address={}.", address);
+        return new AddressManager(project, Arrays.asList(address.split(CommonConfiguration.COMMA)));
     }
 }
