+++ discovery/seata-discovery-servicecomb/src/main/java/io/seata/discovery/registry/servicecomb/client/auth/RBACRequestAuthHeaderProvider.java
@@ -25,8 +25,8 @@ import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.ListeningExecutorService;
 import com.google.common.util.concurrent.MoreExecutors;
-import io.seata.discovery.registry.servicecomb.client.CommonConfiguration;
-import io.seata.discovery.registry.servicecomb.client.EventManager;
+import io.seata.config.servicecomb.client.CommonConfiguration;
+import io.seata.config.servicecomb.client.EventManager;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.servicecomb.foundation.auth.AuthHeaderProvider;
 import org.apache.servicecomb.service.center.client.OperationEvents;
